fixed test issues
